backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
pillow==11.1.0
